moved all player related utility classes to player in entity
