+++ src/main/java/teamport/aether/entity/player/PlayerUntil.java
@@ -1,4 +1,4 @@
-package teamport.aether.helper;
+package teamport.aether.entity.player;
 
 import net.minecraft.core.entity.player.Player;
 import net.minecraft.core.item.ItemStack;
@@ -7,8 +7,8 @@ import teamport.aether.items.AetherItems;
 import static teamport.aether.items.accessory.SlotAccessory.TRINKET_1_SLOT;
 import static teamport.aether.items.accessory.SlotAccessory.TRINKET_2_SLOT;
 
-public class AetherUtil {
-    private AetherUtil(){/* no need to initiate*/}
+public class PlayerUntil {
+    private PlayerUntil(){/* no need to initiate*/}
 
 
     public static boolean isSilkTouch(Player player) {
