+++ src/main/java/teamport/aether/mixin/armor/player/gravitite/MobMixinJump.java
@@ -15,7 +15,7 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import teamport.aether.helper.ContainerHelper;
+import teamport.aether.entity.player.ContainerHelper;
 import teamport.aether.helper.ParticleMaker;
 import teamport.aether.items.AetherArmorMaterial;
