+++ src/main/java/teamport/aether/mixin/armor/player/obsidian/PlayerMixinNoKnockbackFling.java
@@ -5,7 +5,7 @@ import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import net.minecraft.core.entity.Entity;
 import net.minecraft.core.entity.player.Player;
 import org.spongepowered.asm.mixin.Mixin;
-import teamport.aether.helper.ContainerHelper;
+import teamport.aether.entity.player.ContainerHelper;
 import teamport.aether.items.AetherArmorMaterial;
 
 @Mixin(value = Entity.class, remap = false)
